+++ themes/GamePlatform_VelzonSaas/assets/js/pages/home.js
@@ -1,12 +1,4 @@
-require( '../../css/games.css' );
-
 $( function()
 {
-	$( '.btnGame' ).on( 'click', function( e )
-	{
-	   e.preventDefault();
-	   e.stopPropagation();
-	   
-	   document.location   = $( this ).attr( 'data-url' );
-	});
+	
 });